refactor lexer helpers
