+++ src/lexer.c
@@ -14,53 +14,7 @@
 #include "vector.h"
 #include "util.h"
 #include "error.h"
-
-typedef struct {
-    const char *kw;
-    token_type_t tok;
-} keyword_t;
-
-static const keyword_t keyword_table[] = {
-    { "if",       TOK_KW_IF },
-    { "else",     TOK_KW_ELSE },
-    { "do",       TOK_KW_DO },
-    { "while",    TOK_KW_WHILE },
-    { "for",      TOK_KW_FOR },
-    { "break",    TOK_KW_BREAK },
-    { "continue", TOK_KW_CONTINUE },
-    { "goto",     TOK_KW_GOTO },
-    { "switch",   TOK_KW_SWITCH },
-    { "case",     TOK_KW_CASE },
-    { "default",  TOK_KW_DEFAULT },
-    { "sizeof",   TOK_KW_SIZEOF },
-    { "int",      TOK_KW_INT },
-    { "char",     TOK_KW_CHAR },
-    { "float",    TOK_KW_FLOAT },
-    { "double",   TOK_KW_DOUBLE },
-    { "short",    TOK_KW_SHORT },
-    { "long",     TOK_KW_LONG },
-    { "bool",     TOK_KW_BOOL },
-    { "_Bool",    TOK_KW_BOOL },
-    { "_Complex", TOK_KW_COMPLEX },
-    { "alignas", TOK_KW_ALIGNAS },
-    { "_Alignof", TOK_KW_ALIGNOF },
-    { "unsigned", TOK_KW_UNSIGNED },
-    { "void",     TOK_KW_VOID },
-    { "enum",     TOK_KW_ENUM },
-    { "struct",   TOK_KW_STRUCT },
-    { "union",    TOK_KW_UNION },
-    { "typedef",  TOK_KW_TYPEDEF },
-    { "static",   TOK_KW_STATIC },
-    { "extern",   TOK_KW_EXTERN },
-    { "const",    TOK_KW_CONST },
-    { "volatile", TOK_KW_VOLATILE },
-    { "restrict", TOK_KW_RESTRICT },
-    { "register", TOK_KW_REGISTER },
-    { "inline",   TOK_KW_INLINE },
-    { "_Noreturn", TOK_KW_NORETURN },
-    { "_Static_assert", TOK_KW_STATIC_ASSERT },
-    { "return",   TOK_KW_RETURN }
-};
+#include "lexer_internal.h"
 
 typedef struct {
     const char *op;
@@ -97,18 +51,9 @@ static const punct_entry_t punct_table[] = {
 /* Iterate the keyword table and return the matching token type, or
  * TOK_IDENT if the text is not a keyword.
  */
-static token_type_t lookup_keyword(const char *str, size_t len)
-{
-    for (size_t i = 0; i < sizeof(keyword_table) / sizeof(keyword_table[0]); i++) {
-        const keyword_t *kw = &keyword_table[i];
-        if (strlen(kw->kw) == len && strncmp(kw->kw, str, len) == 0)
-            return kw->tok;
-    }
-    return TOK_IDENT;
-}
 
 /* Helper to create and append a token to the vector */
-static void append_token(vector_t *vec, token_type_t type, const char *lexeme,
+void append_token(vector_t *vec, token_type_t type, const char *lexeme,
                          size_t len, size_t line, size_t column)
 {
     char *text = vc_strndup(lexeme, len);
@@ -206,254 +151,6 @@ static void skip_whitespace(const char *src, size_t *i, size_t *line,
     }
 }
 
-/* Read an identifier or keyword starting at src[*i] */
-static void read_identifier(const char *src, size_t *i, size_t *col,
-                            vector_t *tokens, size_t line)
-{
-    size_t start = *i;
-    while (isalnum((unsigned char)src[*i]) || src[*i] == '_')
-        (*i)++;
-    size_t len = *i - start;
-    token_type_t type = TOK_IDENT;
-    if (src[*i] == ':') {
-        (*i)++; /* consume ':' */
-        append_token(tokens, TOK_LABEL, src + start, len, line, *col);
-        *col += len + 1;
-        return;
-    }
-    type = lookup_keyword(src + start, len);
-    append_token(tokens, type, src + start, len, line, *col);
-    *col += len;
-}
-
-/* Parse a numeric literal */
-static void read_number(const char *src, size_t *i, size_t *col,
-                        vector_t *tokens, size_t line)
-{
-    size_t start = *i;
-    int is_float = 0;
-
-    if (src[*i] == '0' && (src[*i + 1] == 'x' || src[*i + 1] == 'X')) {
-        (*i) += 2; /* consume 0x */
-        while (isxdigit((unsigned char)src[*i]))
-            (*i)++;
-    } else if (src[*i] == '0') {
-        (*i)++;
-        while (src[*i] >= '0' && src[*i] <= '7')
-            (*i)++;
-    } else {
-        while (isdigit((unsigned char)src[*i]))
-            (*i)++;
-    }
-
-    if (src[*i] == '.') {
-        is_float = 1;
-        (*i)++; /* consume '.' */
-        while (isdigit((unsigned char)src[*i]))
-            (*i)++;
-    }
-
-    if (src[*i] == 'e' || src[*i] == 'E') {
-        is_float = 1;
-        (*i)++; /* consume exponent marker */
-        if (src[*i] == '+' || src[*i] == '-')
-            (*i)++;
-        while (isdigit((unsigned char)src[*i]))
-            (*i)++;
-    }
-
-    while (src[*i] == 'u' || src[*i] == 'U' ||
-           src[*i] == 'l' || src[*i] == 'L') {
-        (*i)++;
-    }
-
-    token_type_t type = TOK_NUMBER;
-    if (src[*i] == 'i' || src[*i] == 'I') {
-        (*i)++;
-        type = TOK_IMAG_NUMBER;
-    }
-
-    size_t len = *i - start;
-    (void)is_float; /* not used but kept for clarity */
-    append_token(tokens, type, src + start, len, line, *col);
-    *col += len;
-}
-
-/* Simple lookup table for escape sequences */
-typedef struct {
-    char esc;
-    char value;
-} escape_entry_t;
-
-static const escape_entry_t escape_table[] = {
-    { 'n', '\n' },
-    { 't', '\t' },
-    { 'r', '\r' },
-    { 'b', '\b' },
-    { 'f', '\f' },
-    { 'v', '\v' },
-    { '\\', '\\' },
-    { '\'', '\'' },
-    { '"', '"' }
-};
-
-/* Parse up to three octal digits and clamp to 255 */
-static int parse_octal(const char *src, size_t *i,
-                       size_t line, size_t column)
-{
-    int value = 0;
-    int digits = 0;
-    int overflow = 0;
-    while (digits < 3 && src[*i] >= '0' && src[*i] <= '7') {
-        int digit = src[*i] - '0';
-        int next = value * 8 + digit;
-        if (next > 255) {
-            overflow = 1;
-            value = 255;
-        } else {
-            value = next;
-        }
-        (*i)++;
-        digits++;
-    }
-    if (overflow) {
-        error_set(line, column, error_current_file, error_current_function);
-        error_print("Escape sequence out of range");
-    }
-    return value;
-}
-
-/* Parse one or two hexadecimal digits following 'x' */
-static int parse_hex(const char *src, size_t *i)
-{
-    (*i)++; /* skip 'x' */
-    int value = 0;
-    int digits = 0;
-    while (isxdigit((unsigned char)src[*i]) && digits < 2) {
-        char d = src[*i];
-        int hexval = (d >= '0' && d <= '9') ? d - '0' :
-                     (d >= 'a' && d <= 'f') ? d - 'a' + 10 :
-                     (d >= 'A' && d <= 'F') ? d - 'A' + 10 : 0;
-        value = value * 16 + hexval;
-        (*i)++;
-        digits++;
-    }
-    return value;
-}
-
-/* Translate escape sequences within character and string literals */
-static int unescape_char(const char *src, size_t *i,
-                         size_t line, size_t column)
-{
-    if (!src[*i])
-        return 0;
-
-    char c = src[*i];
-    for (size_t n = 0; n < sizeof(escape_table)/sizeof(escape_table[0]); n++) {
-        if (c == escape_table[n].esc) {
-            (*i)++;
-            return escape_table[n].value;
-        }
-    }
-
-    if (c == 'x')
-        return parse_hex(src, i);
-    if (c >= '0' && c <= '7')
-        return parse_octal(src, i, line, column);
-
-    (*i)++;
-    return c;
-}
-
-/* Parse a character constant like '\n' or 'a' */
-static void read_char_const(const char *src, size_t *i, size_t *col,
-                            vector_t *tokens, size_t line,
-                            token_type_t tok_type)
-{
-    size_t column = *col;
-    (*i)++; /* skip opening quote */
-    (*col)++;
-    if (!src[*i]) {
-        error_set(line, column, error_current_file, error_current_function);
-        error_print("Missing closing quote");
-        append_token(tokens, TOK_UNKNOWN, "", 0, line, column);
-        return;
-    }
-
-    unsigned char value = (unsigned char)src[*i];
-    if (value == '\\') {
-        (*i)++; /* skip backslash */
-        value = (unsigned char)unescape_char(src, i, line, column);
-    } else {
-        (*i)++; /* consume character */
-    }
-    (*col)++;
-
-    if (src[*i] != '\'') {
-        error_set(line, column, error_current_file, error_current_function);
-        error_print("Missing closing quote");
-        append_token(tokens, TOK_UNKNOWN, "", 0, line, column);
-        return;
-    }
-
-    (*i)++;
-    (*col)++;
-
-    char buf[2] = {(char)value, '\0'};
-    append_token(tokens, tok_type, buf, 1, line, column);
-}
-
-/* Parse a double-quoted string literal */
-static int read_string_lit(const char *src, size_t *i, size_t *col,
-                           vector_t *tokens, size_t line,
-                           token_type_t tok_type)
-{
-    size_t column = *col;
-    (*i)++; /* skip opening quote */
-    (*col)++;
-
-    vector_t buf_v;
-    vector_init(&buf_v, sizeof(char));
-
-    while (src[*i] && src[*i] != '"') {
-        char c = src[*i];
-        if (c == '\\') {
-            (*i)++; /* skip backslash */
-            c = (char)unescape_char(src, i, line, column);
-        } else {
-            (*i)++; /* consume character */
-        }
-        if (!vector_push(&buf_v, &c)) {
-            vc_oom();
-            vector_free(&buf_v);
-            return 0;
-        }
-        (*col)++;
-    }
-    /* NUL-terminate the buffer for convenience */
-    char nul = '\0';
-    if (!vector_push(&buf_v, &nul)) {
-        vc_oom();
-        vector_free(&buf_v);
-        return 0;
-    }
-    if (src[*i] == '"') {
-        (*i)++;
-        (*col)++;
-
-        append_token(tokens, tok_type, buf_v.data, buf_v.count - 1,
-                     line, column);
-        vector_free(&buf_v);
-        return 1;
-    } else {
-        error_set(line, column, error_current_file, error_current_function);
-        error_print("Missing closing quote");
-        vector_free(&buf_v);
-        append_token(tokens, TOK_UNKNOWN, "", 0, line, column);
-        return 1;
-    }
-}
-
 /* Convert punctuation characters to tokens */
 static void read_punct(char c, vector_t *tokens, size_t line, size_t column)
 {
@@ -486,65 +183,6 @@ static void read_punct(char c, vector_t *tokens, size_t line, size_t column)
     append_token(tokens, type, &c, 1, line, column);
 }
 
-/* Shallow scanning helpers */
-static int scan_identifier(const char *src, size_t *i, size_t *col,
-                           vector_t *tokens, size_t line)
-{
-    char c = src[*i];
-    if (!isalpha((unsigned char)c) && c != '_')
-        return 0;
-    read_identifier(src, i, col, tokens, line);
-    return 1;
-}
-
-static int scan_number(const char *src, size_t *i, size_t *col,
-                       vector_t *tokens, size_t line)
-{
-    if (!isdigit((unsigned char)src[*i]))
-        return 0;
-    read_number(src, i, col, tokens, line);
-    return 1;
-}
-
-static int scan_string(const char *src, size_t *i, size_t *col,
-                       vector_t *tokens, size_t line)
-{
-    if (src[*i] != '"')
-        return 0;
-    if (!read_string_lit(src, i, col, tokens, line, TOK_STRING))
-        return -1;
-    return 1;
-}
-
-static int scan_char(const char *src, size_t *i, size_t *col,
-                     vector_t *tokens, size_t line)
-{
-    if (src[*i] != '\'')
-        return 0;
-    read_char_const(src, i, col, tokens, line, TOK_CHAR);
-    return 1;
-}
-
-static int scan_wstring(const char *src, size_t *i, size_t *col,
-                        vector_t *tokens, size_t line)
-{
-    if (src[*i] != 'L' || src[*i + 1] != '"')
-        return 0;
-    (*i)++; (*col)++;
-    if (!read_string_lit(src, i, col, tokens, line, TOK_WIDE_STRING))
-        return -1;
-    return 1;
-}
-
-static int scan_wchar(const char *src, size_t *i, size_t *col,
-                      vector_t *tokens, size_t line)
-{
-    if (src[*i] != 'L' || src[*i + 1] != '\'')
-        return 0;
-    (*i)++; (*col)++;
-    read_char_const(src, i, col, tokens, line, TOK_WIDE_CHAR);
-    return 1;
-}
 
 static int scan_punct_table(const char *src, size_t *i, size_t *col,
                             vector_t *tokens, size_t line)
